refactor: remove old type files
