+++ packages/slinkity/defineConfig.d.ts
@@ -1,3 +1,3 @@
-import { UserSlinkityConfig } from './cli/types'
+import { UserSlinkityConfig } from './@types'
 
 export function defineConfig(userConfig: Partial<UserSlinkityConfig>): UserSlinkityConfig
