+++ packages/slinkity/index.d.ts
@@ -1,2 +1,2 @@
 export { defineConfig } from './defineConfig'
-export type { UserSlinkityConfig, Hydrate, Renderer } from './cli/types'
+export type { UserSlinkityConfig, Hydrate, Renderer } from './@types'
