+++ packages/slinkity/package.json
@@ -12,8 +12,7 @@
     "index.d.ts",
     "defineConfig.js",
     "defineConfig.d.ts",
-    "cli/types.d.ts",
-    "eleventyConfig/types.d.ts",
+    "@types.ts",
     "client.js"
   ],
   "homepage": "https://slinkity.dev",
